Premiers pas dans la connexion et l'autorisation

+++ Routes/groups.js
@@ -33,7 +33,7 @@ router.get("/day_groups", async (req, res) => {
         const dayGroups = await getDayGroups(association_id, actualDay);
         res.send(dayGroups);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération des groupes du jour.");
     }
 })
@@ -49,7 +49,7 @@ router.get("/day_groups/count", async (req, res) => {
         const dayGroups_count = await getDayGroupsCount(association_id, actualDay);
         res.status(200).json({ day_groups_count: dayGroups_count });
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération des groupes du jour.");
     }
 })
@@ -60,7 +60,7 @@ router.get("/day/:day_id", async (req, res) => {
         const dayGroups = await getDayGroups(association_id, req.params.day_id);
         res.send(dayGroups);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération des groupes de ce jour.");
     }
 })
@@ -73,7 +73,7 @@ router.get("/:group_id", async (req, res) => {
         }
         res.send(group);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération du groupe.");
     }
 })
@@ -87,7 +87,7 @@ router.get("/:group_id/members", async (req, res) => {
         const members = await getGroupMembers(group.id);
         res.send(members);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération des membres du groupe.");
     }
 })
@@ -101,7 +101,7 @@ router.get("/:group_id/users", async (req, res) => {
         const users = await getGroupUsers(group.id);
         res.send(users);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la récupération des membres du groupe.");
     }
 })
@@ -109,13 +109,16 @@ router.get("/:group_id/users", async (req, res) => {
 // POST //
 
 router.post("/", async (req, res) => {
+    if (req.auth.userLvl < 1) {
+        return res.status(403).send("Vous n'avez pas les droits d'accès.");
+    }
     try {
         const { name, group_day, members_max, start_time, end_time } = req.body;
         const association_id = 7
         const group = await createGroup(name, association_id, group_day, members_max, start_time, end_time);
         res.send(`Le groupe ${name} a été créé.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la création du groupe.");
     }
 })
@@ -129,12 +132,15 @@ router.post("/:group_id/members", async (req, res) => {
         });
         res.send(`Les membres ${req.body.members_list.join(", ")} ont été ajouté au groupe ${group_name} avec succès.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de l'ajout des membres au groupe.");
     }
 })
 
 router.post("/:group_id/users", async (req, res) => {
+    if (req.auth.userLvl < 1) {
+        return res.status(403).send("Vous n'avez pas les droits d'accès.");
+    }
     try {
         const group = await getGroup(req.params.group_id, 7);
         const group_name = group.name
@@ -143,7 +149,7 @@ router.post("/:group_id/users", async (req, res) => {
         });
         res.send(`Les users ${req.body.users_list.join(", ")} ont été ajouté au groupe ${group_name} avec succès.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de l'ajout des users au groupe.");
     }
 })
@@ -151,6 +157,9 @@ router.post("/:group_id/users", async (req, res) => {
 // PUT //
 
 router.put("/:group_id", async (req, res) => {
+    if (req.auth.userLvl < 1) {
+        return res.status(403).send("Vous n'avez pas les droits d'accès.");
+    }
     try {
         const { name, group_day, members_max, start_time, end_time } = req.body;
         const id = req.params.group_id
@@ -161,7 +170,7 @@ router.put("/:group_id", async (req, res) => {
         const update_group = await updateGroup(id, name, group_day, members_max, start_time, end_time);
         res.send(`Le groupe avec l'id ${id} a été modifié avec succès.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la modification du groupe.");
     }
 })
@@ -176,7 +185,7 @@ router.put("/:group_id/presence", async (req, res) => {
         await updateGroupPresence(group.id, req.body.members_list);
         res.send(`Les présences du groupe ${group_name} ont été mises à jour.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de l'assignation des présences du groupe");
     }
 })
@@ -184,12 +193,15 @@ router.put("/:group_id/presence", async (req, res) => {
 // DELETE //
 
 router.delete("/:group_id", async (req, res) => {
+    if (req.auth.userLvl < 1) {
+        return res.status(403).send("Vous n'avez pas les droits d'accès.");
+    }
     try {
         const id = req.params.group_id
         const group = await deleteGroup(id);
         res.send(`Le groupe ${group.name} a été supprimé avec succès.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la suppression du groupe.");
     }
 })
@@ -203,12 +215,15 @@ router.delete("/:group_id/members", async (req, res) => {
         });
         res.send(`Les membres ${req.body.members_list.join(", ")} du groupe ${group_name} ont été supprimé.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la suppression des membres du groupe.");
     }
 })
 
 router.delete("/:group_id/users", async (req, res) => {
+    if (req.auth.userLvl < 1) {
+        return res.status(403).send("Vous n'avez pas les droits d'accès.");
+    }
     try {
         const group = await getGroup(req.params.group_id, 7);
         const group_name = group.name
@@ -217,7 +232,7 @@ router.delete("/:group_id/users", async (req, res) => {
         });
         res.send(`Les users ${req.body.users_list.join(", ")} du groupe ${group_name} ont été supprimé.`);
     } catch (error) {
-        console.error(error);
+        
         res.status(500).send("Une erreur est survenue lors de la suppression des users du groupe.");
     }
 })
